README.md: expand credit for original (#43)
* README.md: expand credit for original

Sudo was invented on Berkeley Unix on a VAX, and the current version is maintained by Todd Miller of the OpenBSD Unix project. It is thus correct to refer to the original as Unix/Linux as it is not a project originated by the Linux developers.

* Update README.md again, in ref to original sudo

Suggested by "omni" to have 'less occurrences of "Unix/Linux" in the README' - reads better this way.